integrating model code
