--- tcri/model/model.py
@@ -1,137 +1,0 @@
-import pyro
-import torch
-import pyro.distributions as dist
-from pyro.infer import SVI, Trace_ELBO
-from torch.utils.data import DataLoader, Dataset
-import numpy as np
-import pyro.distributions as dist
-from torch.distributions import constraints
-
-
-class AnnDataTCRDataset(Dataset):
-    def __init__(self, adata, gene_list, tcr_key):
-        self.adata = adata
-        self.gene_list = gene_list
-        self.tcr_key = tcr_key
-
-        self.gene_indices = [np.where(adata.var_names == gene)[0][0] for gene in gene_list]
-        self.gene_expression = self.adata.X[:, self.gene_indices].toarray() if hasattr(self.adata.X, "toarray") else self.adata.X[:, self.gene_indices]
-        self.tcrs = self.adata.obs[self.tcr_key].astype('category').cat.codes.values
-        self.unique_tcrs = self.adata.obs[self.tcr_key].astype('category').cat.categories
-
-    def __len__(self):
-        return self.gene_expression.shape[0]
-
-    def __getitem__(self, idx):
-        x = torch.tensor(self.gene_expression[idx], dtype=torch.float32)  # Gene expression
-        z = torch.tensor(self.tcrs[idx], dtype=torch.long)  # TCR
-        return x, z
-
-    def get_tcr_mapping(self):
-        return dict(enumerate(self.unique_tcrs))
-
-def model(genes, tcrs, K, D):
-    """
-    Args:
-        genes: Batch of gene expression data (batch_size, D)
-        tcrs: Batch of TCR indices (batch_size)
-        K: Number of unique TCRs
-        D: Number of genes
-    """
-    probs = pyro.sample("probs", dist.Dirichlet(torch.ones(K, device=genes.device)))  # Shape: (K,)
-
-    with pyro.plate("tcr_plate", K, dim=-2):  # Plate over K (TCRs)
-        with pyro.plate("gene_plate", D, dim=-1):  # Plate over D (genes)
-            mu = pyro.sample(
-                "mu", 
-                dist.LogNormal(
-                    torch.zeros((K, D), device=genes.device),  # Shape: (K, D)
-                    torch.ones((K, D), device=genes.device) * 1.0  # Shape: (K, D)
-                )
-            )  # Shape: (K, D)
-            theta = pyro.sample(
-                "theta", 
-                dist.Gamma(
-                    torch.ones((K, D), device=genes.device) * 2.0,  # Shape: (K, D)
-                    torch.ones((K, D), device=genes.device) * 0.5  # Shape: (K, D)
-                )
-            )  # Shape: (K, D)
-            gate = pyro.sample(
-                "gate", 
-                dist.Beta(
-                    torch.ones((K, D), device=genes.device) * 0.8, 
-                    torch.ones((K, D), device=genes.device) * 0.8
-                )
-            )  # Shape: (K, D)
-
-    with pyro.plate("data", size=genes.shape[0], dim=-2):  # Batch over data (batch_size)
-        z = pyro.sample("z", dist.Categorical(probs), obs=tcrs)  # Shape: (batch_size,)
-
-        mu_z = mu[z]  # Shape: (batch_size, D)
-        theta_z = theta[z]  # Shape: (batch_size, D)
-        gate_z = gate[z]  # Shape: (batch_size, D)
-
-        probs = torch.sigmoid(mu_z / (mu_z + theta_z))  # Ensure this is within [0, 1]
-        
-        zinb = dist.ZeroInflatedNegativeBinomial(
-            total_count=theta_z,
-            probs=probs, 
-            gate=gate_z  
-        )
-        pyro.sample("genes", zinb.to_event(1), obs=genes)  
-
-
-
-def guide(genes, tcrs, K, D):
-
-    probs_q = pyro.param("probs_q", torch.ones(K, device=genes.device), constraint=constraints.simplex)
-    pyro.sample("probs", dist.Dirichlet(probs_q))  # Shape: (K,)
-
-
-    with pyro.plate("tcr_plate", K, dim=-2):  # Plate over K (TCRs)
-        with pyro.plate("gene_plate", D, dim=-1):  # Plate over D (genes)
-            mu_q = pyro.param("mu_q", torch.zeros((K, D), device=genes.device))  # (K, D)
-            sigma_q = pyro.param("sigma_q", torch.ones((K, D), device=genes.device), constraint=constraints.positive)  # Shape: (K, D)
-            pyro.sample("mu", dist.LogNormal(mu_q, sigma_q))  #(K, D)
-
-            theta_q = pyro.param("theta_q", torch.ones((K, D), device=genes.device), constraint=constraints.positive)  # Shape: (K, D)
-            pyro.sample("theta", dist.Gamma(theta_q, 1.0))  # (K, D)
-
-            gate_alpha_q = pyro.param("gate_alpha_q", torch.ones((K, D), device=genes.device), constraint=constraints.positive)  # Shape: (K, D)
-            gate_beta_q = pyro.param("gate_beta_q", torch.ones((K, D), device=genes.device), constraint=constraints.positive)  # Shape: (K, D)
-            pyro.sample("gate", dist.Beta(gate_alpha_q, gate_beta_q))  # (K, D)
-    
-    with pyro.plate("data", size=genes.shape[0], dim=-2):  
-        z_logits = pyro.param("z_logits", torch.randn((genes.shape[0], K), device=genes.device))  # Shape: (batch_size, K)
-        pyro.sample("z", dist.Categorical(logits=z_logits), infer={"is_auxiliary": True})  # Shape: (batch_size,)
-
-def train_model(adata, gene_list, tcr_key, num_epochs=1000, batch_size=10000):
-    losses = []
-    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-    print(f"Using device: {device}")
-    dataset = AnnDataTCRDataset(adata, gene_list, tcr_key)
-    data_loader = DataLoader(dataset, batch_size=batch_size, shuffle=True, drop_last=True)
-
-    K = len(dataset.get_tcr_mapping())  # Number of TCRs
-    D = len(gene_list)  # Number of genes
-
-    pyro.clear_param_store()
-    svi = SVI(
-        model=lambda genes, tcrs, K, D: model(genes, tcrs, K, D), 
-        guide=lambda genes, tcrs, K, D: guide(genes, tcrs, K, D), 
-        optim=pyro.optim.Adam({"lr": 0.01}), 
-        loss=Trace_ELBO()
-    )
-
-    for epoch in range(num_epochs):
-        epoch_loss = 0
-        for batch_idx, (gene_batch, tcr_batch) in enumerate(data_loader):
-            gene_batch = gene_batch.to(device)
-            tcr_batch = tcr_batch.to(device)
-            loss = svi.step(gene_batch, tcr_batch, K=K, D=D)
-            epoch_loss += loss
-            losses.append(loss)
-
-        if epoch % 5 == 0:
-            print(f"Epoch {epoch} Loss: {epoch_loss / len(data_loader):.4f}")
-    return losses
